tests/wpt/web-platform-tests/html/rendering/non-replaced-elements/the-page/iframe-marginwidth-marginheight.html

<!doctype html>
<title>iframe marginwidth and marginheight</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe src="/common/blank.html" marginwidth=0 marginheight=0></iframe>
<script>
onload = () => {
  assert_equals(window[0].document.body.attributes.length, 0, "Number of attributes on the child document's body");
  done();
}
</script>
