LayoutTests/imported/w3c/web-platform-tests/html/rendering/non-replaced-elements/the-page/iframe-marginwidth-marginheight-expected.txt

CONSOLE MESSAGE: line 2659: Error: assert_equals: Number of attributes on the child document's body expected 0 but got 2


FAIL iframe marginwidth and marginheight Error: assert_equals: Number of attributes on the child document's body expected 0 but got 2
